🐛 Workaround to fix issues with Menu and SSR
